chambre client  facture add CrossOrigin annotation

+++ src/main/java/com/pfa/reservationhoteliere/repository/IClientRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.pfa.reservationhoteliere.entity.Client;
 
 
-public interface IClientRepository extends JpaRepository<Client,Integer>{
-	Client findById(int id);
+public interface IClientRepository extends JpaRepository<Client,Long>{
+	Client findById(long id);
 	Client findByReference(String reference);
 }
